+++ src/QueryCat.Backend/Commands/Select/KeyConditionValue/IKeyConditionSingleValueGenerator.cs
@@ -13,6 +13,7 @@ internal interface IKeyConditionSingleValueGenerator
     /// Get the current value.
     /// </summary>
     /// <param name="thread">Execution thread.</param>
-    /// <returns>Value.</returns>
-    VariantValue Get(IExecutionThread thread);
+    /// <param name="value">Result value.</param>
+    /// <returns><c>True</c> if able to return, <c>false</c> otherwise.</returns>
+    bool TryGet(IExecutionThread thread, out VariantValue value);
 }