+++ src/QueryCat.Backend/Storage/FuncUnitRowsInputColumn.cs
@@ -1,3 +1,4 @@
+using QueryCat.Backend.Core;
 using QueryCat.Backend.Core.Data;
 using QueryCat.Backend.Core.Execution;
 using QueryCat.Backend.Core.Types;
@@ -12,7 +13,13 @@ internal sealed class FuncUnitRowsInputColumn(IRowsInput rowsInput, int columnIn
     /// <inheritdoc />
     public ValueTask<VariantValue> InvokeAsync(IExecutionThread thread, CancellationToken cancellationToken = default)
     {
-        rowsInput.ReadValue(columnIndex, out VariantValue value);
+        TryInvokeAsync(thread, out var value, cancellationToken);
         return ValueTask.FromResult(value);
+    }
+
+    internal ValueTask<bool> TryInvokeAsync(IExecutionThread thread, out VariantValue value, CancellationToken cancellationToken = default)
+    {
+        var errorCode = rowsInput.ReadValue(columnIndex, out value);
+        return ValueTask.FromResult(errorCode == ErrorCode.OK);
     }
 }