+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Initial async functions support.
+- Add new Unset method for rows input keys.
 
 ## [0.9.0] - 2025-01-07